Moved isNumeric check to it's own class in CheckGuess package.
